remove go.sum temporarily
It seems that how to calculate hashes is depending on Go
toolchain toolchain version. It means that this library may not be
installed due to 'broken hash' with different Go toolchain from mine. I
believe it is temporal since Go module feature is actively developed.

I decided to remove `go.sum` temporarily. It should be restored in the
future.